Removing unused packaging code and update README
This repo was originally designed for something completely different
than it's used for today. It would have been incredibly misleading for
anyone discovering the repo without any context. This PR is to
completely gut all of the unused packaging code and make targets, as
well as revamp the README entirely to reflect what this collector is
actually used for today.

in progress