Removing unused packaging code and update README (#157)
* Removing unused packaging code and update README

This repo was originally designed for something completely different
than it's used for today. It would have been incredibly misleading for
anyone discovering the repo without any context. This PR is to
completely gut all of the unused packaging code and make targets, as
well as revamp the README entirely to reflect what this collector is
actually used for today.

in progress

* Reworked Makefile and tools, added guides

Reworked the Makefile and Tools setup. This also involved restoring
googet and tar packaging to maintain compatibility with running Kokoro
jobs. Also added new development guides to keep the README smaller.

* go workspaces won't work for this repo

* okay I'll keep the presubmit name

* okay fine I'll put that config file back

* Addressed Makefile comments, added tools guide

The tools setup is pretty unintuitive, so I've added a guide for how
they work to `docs/dev`.

* presubmit: run full presubmit with GPU support

All the checks in presubmit actually should be run with GPU support, as
the linter should be leveraging the build tags to lint the packages
properly.

* dcgmreceiver: use new method of random source

* I disagree with the linter dinging me for this

* Streamline updating opentelemetry

Combine most steps of updating OpenTelemetry into a single make target.

* Collapse make target details

* Update docs/dev/upgrade-opentelemetry.md

Co-authored-by: igorpeshansky <igorpeshansky@users.noreply.github.com>

---------

Co-authored-by: igorpeshansky <igorpeshansky@users.noreply.github.com>