Change to support this repo in a Go workspace
This change allows for this repo to be used as part of a Go workspace.
This is very useful in tandem with either `opentelemetry-operations-go`
or `opentelemetry-collector-contrib` because it allows us to create a
workspace with this collector and the library we want to test, add a
manual replacement in the workspace for that dependency, and be able to
test the collector with local copies of the dependency without changing
anything.

Related: https://github.com/GoogleCloudPlatform/opentelemetry-operations-go/pull/793