Restructure repository for OTel version update
The structure for components in this repo had a fundamental issue. Once
we introduced components that were specific to Google-Built
OpenTelemetry Collector, we had the issue of otelopscol components
targeting an older version of OpenTelemetry. We don't want to force all
components to be on the same OTel version; when targeting different
distributions, the components need to be at whichever version that
distribution is on.

This PR restructures the component directories to separate them into two
separate directories, one for each distribution the component could be
targeting. The Makefile structure is adjusted around this fact. The new
process for upating OTel components for a distribution is to go into the
Makefile for that distribution's components folder and update the
version variables, then in the main directory call the associated target
to update those components.

This PR also does the OTel version updates for Google-Built
OpenTelemetry Collector to prove that the restructure worked
effectively.

Updates were needed to the client auth extension based on:
https://github.com/open-telemetry/opentelemetry-collector-contrib/pull/38451
https://github.com/open-telemetry/opentelemetry-collector-contrib/pull/38601