cmd/distrogen: add spec go_proxy field
Added a field to the spec called `go_proxy` that can be set to allow for
a default `GOPROXY` value to be used in collector binary builds. This
also introduces a new `USE_GO_PROXY` make variable that can be set to
empty to circumvent the configured `go_proxy` if needed.