Bump mapstructure/v2
This PR runs the update to mapstructure now that all modules are
properly included, and found that the only missing bump from the
previous attempt was internal/tools. This will close the security issue
opened against the repository (we weren't at risk but it will be nice to
make it go away).