Add Windows services
This adds Windows Services for GBOC when running in package mode.
This was tested offline by doing a package build, uploading the
package to a private bucket, and running the smoke tests in Kokoro
using that bucket.

This also renames the package on Windows from
google-built-opentelemetry-collector to otelcol-google so that
it's consistent with Linux.